Enable cops that were previously checked by pre-commit hook and CI
These checks were run by the pre-commmit hook and in CI, but they were
disabled in `.rubocop.yml`. The reason that they were run anyway was
that we previously had a separate hard-coded list of checks in
`util/hooks/pre_commit_rubocop_config.yml`, but we're getting rid of it
to simplify things.